refactor(dashboard): split MemoryOptimizationCenter god component (#80, part 1)
Resolves smell GM-7. The page was 1,269 LOC owning 25+
useState/useEffect/useRef and 17 inline handlers.

Split:
- `MemoryOptimizationCenter.tsx` 1269 → **365 LOC** (orchestrator only)
- `hooks/useMemoryOptimization.ts` (NEW, 442 LOC) — owns the
  abort-controller + processing state-machine for the long-running
  optimization flow (fetchSuggestions/applySelected/abort/ETA tracking)
- `components/OptimizationFilters.tsx` (NEW, 81 LOC) — agent dropdown,
  date range, priority controls
- `components/SuggestionList.tsx` (NEW, 264 LOC) — checkbox list +
  per-row actions
- `components/OptimizationDetailModal.tsx` (NEW, 236 LOC) — suggestion
  detail viewer (lifted from inline JSX)
- `components/OptimizationDebugPanel.tsx` (NEW, 169 LOC) — block ID
  debug probe (lifted from inline section)

All sub-files ≤ 500 LOC. Default export of `MemoryOptimizationCenter.tsx`
preserved → no caller-side changes (App.tsx import path unchanged).

Design call: kept `selectedSuggestion`/`showDetailModal` and
`selectedBlockId`/`showBlockDetailModal` as orchestrator-local state
because both modal pairs are pure UI orchestration with no business
side-effects; threading two extra setters through the hook's return
type would have been gratuitous coupling.

Verification:
- bun run test: **294/294 passed** (49 suites)
- bun run build: clean, 2.34s

Issue #80 has two parts; OrganizationManagement split follows in a
separate commit.
